refactor: dto 이름 변경 및 패키지 구조 수정
