+++ src/main/java/com/project/baedalsodae/menu/dto/requestDto/item/MenuItemPatchRequestDto.java
@@ -1,13 +1,12 @@
-package com.project.baedalsodae.menu.dto.requestDto;
+package com.project.baedalsodae.menu.dto.requestDto.item;
 
 import com.project.baedalsodae.menu.entity.enums.MenuStatus;
 import java.util.UUID;
 
-public record MenuPatchRequestDto(
+public record MenuItemPatchRequestDto(
     String name,
     String description,
     Integer price,
-    Integer orderNo,
     Boolean isPopular,
     UUID categoryId,
     MenuStatus menuStatus) {}
