+++ src/main/java/com/project/baedalsodae/menu/service/MenuItemService.java
@@ -1,15 +1,15 @@
 package com.project.baedalsodae.menu.service;
 
-import com.project.baedalsodae.menu.dto.requestDto.MenuPatchRequestDto;
-import com.project.baedalsodae.menu.dto.requestDto.MenuPutRequestDto;
-import com.project.baedalsodae.menu.dto.responseDto.MenuItemResponseDto;
+import com.project.baedalsodae.menu.dto.requestDto.item.MenuItemPatchRequestDto;
+import com.project.baedalsodae.menu.dto.requestDto.item.MenuItemPutRequestDto;
+import com.project.baedalsodae.menu.dto.responseDto.item.MenuItemResponseDto;
 import java.util.UUID;
 
 public interface MenuItemService {
 
-  MenuItemResponseDto updateMenuItem(UUID menuItemId, MenuPutRequestDto request);
+  MenuItemResponseDto updateMenuItem(UUID menuItemId, MenuItemPutRequestDto request);
 
-  MenuItemResponseDto patchMenuItem(UUID menuItemId, MenuPatchRequestDto request);
+  MenuItemResponseDto patchMenuItem(UUID menuItemId, MenuItemPatchRequestDto request);
 
   void deleteMenuItem(UUID menuItemId);
 }
